Find Deletion Distance beween two Strings
Determine the deletion distance between two strings? Let's define the deletion distance as the numbers of characters needed to delete from two strings to make them equal.
For example, the deletion distance of algo and daily is 5. The reason is we can delete the go (2 deletion) in algo, and the d, i and y (3 deletions) in daily.