Implement Atoi
Your task  is to implement the function atoi. The function takes a string(str) as argument and converts it to an integer and returns it.

Note: You are not allowed to use inbuilt function.

Input:
str = 123
Output: 123

Input:
str = 21a
Output: -1
Explanation: Output is -1 as all characters are not digit only.